refactor: separate emergency diagnostic from chat area
